requirements.txt
google-api-python-client==2.120.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests==2.32.4
plyer==2.1.0
python-dotenv==1.0.1
langgraph
langchain
schedule